Update Addon SDK v2 for next stable release
Update all plugins and behaviors to explicitly use modules via adding main.js/.ts and a call to SetRuntimeModuleMainScript(). Also update code to use new runtime.sdk namespace to avoid using a deprecated method. This increases the minimum Construct version requirement to r401+, so this is now specified as a requirement in addon.json. Note: at the time of this commit, these releases are only available on the beta channel; the next stable should arrive in the next few weeks with full support for this update to the SDK.